+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Implemented enhancements
 
+- GP2-1270 - Imported Income data
 - GP2-1218 - Country data alignment
 - GP2-173 - Updated Male population CSV
 - GP2-1068 - Apply black and isort autoformatting to codebase incl makefile additions